complete 02.Inversions

+++ 02.Inversions/include/MergeSort.hpp
@@ -12,7 +12,7 @@ namespace MergeSort {
     void merge(const I first_begin, const I first_end,
                const I second_begin, const I second_end,
                I merge_space_begin,
-               int* inversion_count, const Compare comp) {
+               unsigned long long* inversion_count, const Compare comp) {
         I lower = first_begin, upper = second_begin, merge = merge_space_begin;
         while (lower != first_end && upper != second_end) {
             if (comp(*lower, *upper)) {
@@ -29,7 +29,7 @@ namespace MergeSort {
     template<typename I, typename Compare>
     void sortHelper(const I begin, const I end,
                     const I merge_space_begin,
-                    int *inversion_count,
+                    unsigned long long *inversion_count,
                     const Compare comp) {
         auto n = end - begin;
         if (n <= 1) return;
@@ -45,16 +45,16 @@ namespace MergeSort {
 
     template <typename I, typename Compare = std::less< typename I::value_type> >
     void sort(const I begin, const I end,
-              int* inversion_count = nullptr,
+              unsigned long long* inversion_count = nullptr,
               const Compare comp = Compare()) {
         std::vector<typename I::value_type> merge_space(end - begin); 
         sortHelper(begin, end, std::begin(merge_space), inversion_count, comp);
     }
 
     template<typename I, typename Compare = std::less< typename I::value_type> >
-    int sortAndGetInversionCount(const I begin, const I end,
+    unsigned long long sortAndGetInversionCount(const I begin, const I end,
                                  const Compare comp = Compare()) {
-        int inversion_count = 0;
+        unsigned long long inversion_count = 0;
         sort(begin, end, &inversion_count, comp);
         return inversion_count;
     }
